requirements.txt
django=4.2.15
djangorestframework=3.15.2
django-filter=24.3
Markdown=3.7
djangorestframework-simplejwt=5.3.1